Documentation
Hello,

While completing this artwork, I used classes to construct the satellite as an object that can change the orbital radius upon changing the argument. I made a planet with a ring, alongside stars and a black background to create space. I used a rectangle, shredded to form an ellipse for the satellite with only one constructor: orbital radius. Additionally, I created a function containing the code to construct a round, elliptical satellite. 

It was quite confusing for me to keep track of all variables and constructors. I still fail to understand the utility and usefulness of classes. I added it primarily as an assignment parameter. I'm confident I could have created the same artwork without classes.

In the upcoming week, I wish to explore classes more to harness their utmost potential.

Best,
Syed Fahad Rizwan